+++ src/main/java/ajc/formation/spring/bibliotheque/repositories/UtilisateurRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ajc.formation.spring.bibliotheque.entities.Utilisateur;
 
 
-public interface UtilisateurRepository extends JpaRepository<Utilisateur, Long>{
+public interface UtilisateurRepository extends JpaRepository<Utilisateur, Integer>{
 	Optional<Utilisateur> findByLogin(String login);
 
 }
